feat: add user autentication

+++ src/pages/api/categories/[id].ts
@@ -1,13 +1,26 @@
 import { NextApiRequest, NextApiResponse } from 'next'
 import { prisma } from '../../../services/prisma'
+import { getSession } from 'next-auth/react'
 
 export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse,
 ) {
   if (req.method === 'GET') {
+    const session = await getSession({ req })
+    const user = await prisma.user.findUnique({
+      where: {
+        email: session?.user?.email ?? '',
+      },
+    })
+
+    if (!user) {
+      return res.status(400).json({
+        message: 'User not found!',
+      })
+    }
     const id = String(req.query.id)
-    const category = await prisma.category.findUniqueOrThrow({
+    const category = await prisma.category.findUnique({
       where: {
         id,
       },
@@ -15,6 +28,19 @@ export default async function handler(
 
     return res.status(200).json(category)
   } else if (req.method === 'PUT') {
+    const session = await getSession({ req })
+    const user = await prisma.user.findUnique({
+      where: {
+        email: session?.user?.email ?? '',
+      },
+    })
+
+    if (!user) {
+      return res.status(400).json({
+        message: 'User not found!',
+      })
+    }
+
     const id = String(req.query.id)
     const { amount, percentage, title } = req.body
     const category = await prisma.category.update({
@@ -25,11 +51,29 @@ export default async function handler(
         amount,
         percentage,
         title,
+        user: {
+          connect: {
+            email: user.email ?? '',
+          },
+        },
       },
     })
 
     return res.status(201).json(category)
   } else if (req.method === 'DELETE') {
+    const session = await getSession({ req })
+    const user = await prisma.user.findUnique({
+      where: {
+        email: session?.user?.email ?? '',
+      },
+    })
+
+    if (!user) {
+      return res.status(400).json({
+        message: 'User not found!',
+      })
+    }
+
     const id = String(req.query.id)
     await prisma.category.delete({
       where: {
